pyproject.toml
[project]
name = "vibe-coding-homeworks-dk-1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "intersystems-irispython",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy-iris>=0.19.0",
]

[tool.uv.sources]
intersystems-irispython = { path = "../../../../../opt/isc/iris/dev/python/intersystems_irispython-5.3.2-cp38.cp39.cp310.cp311.cp312.cp313.cp314-cp38.cp39.cp310.cp311.cp312.cp313.cp314-manylinux2014_x86_64.manylinux_2_17_x86_64.whl" }
